Replace acme client to interface - for mocking
